+++ test/behave.ts
@@ -50,7 +50,18 @@ function make(name: string, delay: number) {
 
 function unlink() {
 	try {
-		fs.unlinkSync(testDb);
+		if (dbExists())
+			fs.unlinkSync(testDb);
 	} catch (ex) {
+		throw new Error("Database exists, but failed to delete");
+	}
+}
+
+function dbExists() {
+	try {
+		fs.statSync(testDb);
+		return true;
+	} catch (ex) {
+		return false;
 	}
 }
